# specs/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Service API
  version: 1.0.0
  description: |
    Template-based service API (Template v3.3.8).

    **Template Core Endpoints** (Keep these):
    - /health - Health check (AC-TPL-001)
    - /version - Build information (AC-TPL-002)
    - /metrics - Prometheus metrics (AC-TPL-007)
    - Error envelope - Standardized errors (AC-TPL-003, AC-TPL-004)

    **Platform Introspection Endpoints** (Governance visibility):
    - /platform/status - Governance health, ledger counts, policy status
    - /platform/docs/index - Documentation inventory with validation
    - /platform/ui/contract - Governed UI contract (screens, regions, data-uiid anchors)
    - /platform/friction - Development friction log (DevEx issues)
    - /platform/questions - Design questions and ambiguities
    - /platform/forks - Template forks/branches
    - /platform/agent/hints - Prioritized work suggestions for AI agents

  x-template-version: "3.3.8"
  x-schema-version: "1.0"
  x-last-updated: "2025-11-17"
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: http://localhost:8080
    description: Local development server
security:
  - ApiKeyAuth: []
paths:
...
  /platform/status:
    get:
      summary: Platform governance status
      description: |
        Returns comprehensive governance status including service metadata,
        ledger counts, DevEx metrics, friction tracking, and policy status.

        This is the primary endpoint for Backstage and other developer portal
        integrations to understand governance health.
      operationId: getPlatformStatus
      tags: [platform]
      security: []  # Platform endpoints may have separate auth
      responses:
        "200":
          description: Platform status retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlatformStatus'
        "500":
          description: Internal server error (e.g., failed to load spec ledger)
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Failed to load governance data"
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
...
    PlatformStatus:
      type: object
      description: "Complete platform governance status"
      required: [service, governance, errors]
      properties:
        service:
          $ref: '#/components/schemas/ServiceInfo'
        governance:
          $ref: '#/components/schemas/GovernanceInfo'
        config:
          description: "Runtime configuration (may be absent if no config loaded)"
          $ref: '#/components/schemas/ConfigInfo'
        errors:
          description: "Error tracking summary (AC-TPL-ERROR-MAPPING)"
          $ref: '#/components/schemas/ErrorSummary'
    ServiceInfo:
      type: object
      description: "Service metadata"
      required: [service_id, template_version, links, tags]
      properties:
        service_id:
          type: string
          description: "Unique service identifier"
        template_version:
          type: string
          description: "Template version used to create this service"
        display_name:
          type: string
          description: "Human-readable service name (optional)"
        description:
          type: string
          description: "Service description (optional)"
        links:
          type: object
          additionalProperties:
            type: string
          description: "Related links (repo, docs, etc.)"
        tags:
          type: array
          items:
            type: string
          description: "Service tags for categorization"
    GovernanceInfo:
      type: object
      description: "Full governance health metrics"
      required: [ledger, devex, docs, tasks, questions, friction, forks, policies]
      properties:
        ledger:
          $ref: '#/components/schemas/LedgerCounts'
        devex:
          $ref: '#/components/schemas/DevExCounts'
        docs:
          $ref: '#/components/schemas/DocsCounts'
        tasks:
          $ref: '#/components/schemas/TasksCounts'
        questions:
          $ref: '#/components/schemas/QuestionsInfo'
        friction:
          $ref: '#/components/schemas/FrictionInfo'
        forks:
          $ref: '#/components/schemas/ForksInfo'
        policies:
          $ref: '#/components/schemas/PoliciesInfo'
    LedgerCounts:
      type: object
      description: "Governance ledger counts"
      required: [stories, requirements, acs]
      properties:
        stories:
          type: integer
          minimum: 0
          description: "Number of user stories"
        requirements:
          type: integer
          minimum: 0
          description: "Number of requirements"
        acs:
          type: integer
          minimum: 0
          description: "Number of acceptance criteria"
    DevExCounts:
      type: object
      description: "Developer experience metrics"
      required: [commands, flows]
      properties:
        commands:
          type: integer
          minimum: 0
          description: "Number of xtask commands"
        flows:
          type: integer
          minimum: 0
          description: "Number of DevEx flows"
    DocsCounts:
      type: object
      description: "Documentation summary"
      required: [total, design, doc_type_issues]
      properties:
        total:
          type: integer
          minimum: 0
          description: "Total number of documents"
        design:
          type: integer
          minimum: 0
          description: "Number of design documents"
        doc_type_issues:
          type: integer
          minimum: 0
          description: "Documents with doc-type validation issues"
    TasksCounts:
      type: object
      description: "Task counts"
      required: [total]
      properties:
        total:
          type: integer
          minimum: 0
          description: "Total number of tasks"
    QuestionsInfo:
      type: object
      description: "Question tracking"
      required: [open, answered, resolved, total, top_open]
      properties:
        open:
          type: integer
          minimum: 0
          description: "Open questions"
        answered:
          type: integer
          minimum: 0
          description: "Answered questions"
        resolved:
          type: integer
          minimum: 0
          description: "Resolved questions"
        total:
          type: integer
          minimum: 0
          description: "Total questions"
        top_open:
          type: array
          items:
            $ref: '#/components/schemas/QuestionSummary'
          description: "Top open questions (max 3)"
    QuestionSummary:
      type: object
      description: "Question summary entry"
      required: [id, summary, flow]
      properties:
        id:
          type: string
          description: "Question ID"
        summary:
          type: string
          description: "Question summary"
        flow:
          type: string
          description: "Associated DevEx flow"
    FrictionInfo:
      type: object
      description: "Friction log tracking"
      required: [total, open, by_severity, recent]
      properties:
        total:
          type: integer
          minimum: 0
          description: "Total friction entries"
        open:
          type: integer
          minimum: 0
          description: "Open friction issues"
        by_severity:
          $ref: '#/components/schemas/SeverityCounts'
        recent:
          type: array
          items:
            $ref: '#/components/schemas/FrictionSummary'
          description: "Recent friction entries (max 5)"
    SeverityCounts:
      type: object
      description: "Friction counts by severity"
      required: [low, medium, high, critical]
      properties:
        low:
          type: integer
          minimum: 0
        medium:
          type: integer
          minimum: 0
        high:
          type: integer
          minimum: 0
        critical:
          type: integer
          minimum: 0
    FrictionSummary:
      type: object
      description: "Friction log entry summary"
      required: [id, date, severity, summary, category]
      properties:
        id:
          type: string
          description: "Friction entry ID"
        date:
          type: string
          description: "Date of friction entry"
        severity:
          type: string
          enum: [low, medium, high, critical]
          description: "Severity level"
        summary:
          type: string
          description: "Brief description"
        category:
          type: string
          description: "Friction category"
    ForksInfo:
      type: object
      description: "Fork tracking"
      required: [total, ids]
      properties:
        total:
          type: integer
          minimum: 0
          description: "Number of forks"
        ids:
          type: array
          items:
            type: string
          description: "Fork IDs"
    PoliciesInfo:
      type: object
      description: "Policy enforcement status"
      required: [status]
      properties:
        status:
          type: string
          enum: [pass, fail, unknown]
          description: "Overall policy status ('unknown' if not yet evaluated)"
    ConfigInfo:
      type: object
      description: "Runtime configuration summary"
      required: [http_port, settings, secrets_redacted, auth]
      properties:
        env:
          type: string
          description: "Environment name (optional)"
        http_port:
          type: integer
          description: "HTTP server port"
        settings:
          type: object
          additionalProperties: true
          description: "Non-secret configuration settings"
        secrets_redacted:
          type: object
          additionalProperties:
            type: string
          description: "Secret keys with redacted values"
        auth:
          $ref: '#/components/schemas/AuthConfig'
    AuthConfig:
      type: object
      description: "Authentication configuration"
      required: [mode, token_present]
      properties:
        mode:
          type: string
          enum: [open, token, oidc]
          description: "Authentication mode"
        token_present:
          type: boolean
          description: "Whether an auth token is configured"

    # ==========================================================================
    # Error Tracking Schemas (AC-TPL-ERROR-MAPPING)
    # ==========================================================================
    ErrorSummary:
      type: object
      description: "Error tracking summary surfaced via /platform/status"
      required: [has_recent_errors, stats]
      properties:
        has_recent_errors:
          type: boolean
          description: "Whether any errors have occurred since service start"
        last_error:
          description: "The most recent error (if any)"
          $ref: '#/components/schemas/LastErrorSummary'
        stats:
          description: "Aggregated error statistics"
          $ref: '#/components/schemas/ErrorStats'
    LastErrorSummary:
      type: object
      description: "Summary of the last error that occurred"
      required: [category, message, status_code, occurred_at]
      properties:
        category:
          type: string
          description: "Error category (e.g., 'task_not_found', 'invalid_transition', 'internal')"
          example: "task_not_found"
        message:
          type: string
          description: "Human-readable error message"
          example: "Task TASK-TPL-DEMO-001 not found"
        status_code:
          type: integer
          format: int32
          minimum: 400
          maximum: 599
          description: "HTTP status code returned"
          example: 404
        occurred_at:
          type: string
          format: date-time
          description: "When the error occurred (ISO 8601)"
        request_id:
          type: string
          description: "Request ID for correlation (X-Request-ID header value)"
    ErrorStats:
      type: object
      description: "Aggregated error statistics since service start"
      required: [total_errors, client_errors, server_errors]
      properties:
        total_errors:
          type: integer
          format: int64
          minimum: 0
          description: "Total number of errors since service start"
        client_errors:
          type: integer
          format: int64
          minimum: 0
          description: "Number of 4xx client errors"
        server_errors:
          type: integer
          format: int64
          minimum: 0
          description: "Number of 5xx server errors"

    # ==========================================================================
    # Documentation Index Schemas
    # ==========================================================================
...